Add processing date to model
Add processing date to model that allows us for deletion of files after a correct interval
